Creación del controlador y verificaciones en serivicio de citas, cambio de tipo de dato de idusuario en cliente

+++ src/main/java/huellitassoft_web/huellitasoft/dto/client/ClientResponseDTO.java
@@ -75,5 +75,5 @@ public class ClientResponseDTO {
             description = "ID del usuario asociado al cliente",
             example = "5"
     )
-    private Integer idUsuario;
+    private Long idUsuario;
 }
